build: Do not overwrite lv_conf.defaults in source path
When a non-default lv_conf.defaults is used it is currenly copied out of
the configs/ directory replacing the existing lv_conf.defaults file. This
makes changes to source files during build which should be avoided, only
modifications should happen in the build directory.

There is no need to do this as we can simply update the path used to
locate lv_conf.defaults in CMake to the non-default file path and use
it just the same.

Signed-off-by: Andrew Davis <afd@ti.com>